pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-data-pipeline"
version = "1.0.0"
description = "A comprehensive ETL pipeline for YouTube data analytics"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["youtube", "api", "etl", "pipeline", "analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.28.0",
    "psycopg2-binary>=2.9.0",
    "PyYAML>=6.0",
    "python-dotenv>=0.19.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "responses>=0.22.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.991",
    "isort>=5.0.0",
    "pre-commit>=2.0.0",
]
scheduling = [
    "schedule>=1.1.0",
    "APScheduler>=3.9.0",
]
monitoring = [
    "prometheus-client>=0.15.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/youtube-data-pipeline"
Documentation = "https://github.com/yourusername/youtube-data-pipeline/wiki"
Repository = "https://github.com/yourusername/youtube-data-pipeline"
Issues = "https://github.com/yourusername/youtube-data-pipeline/issues"

[project.scripts]
youtube-pipeline = "src.scripts.run_pipeline:main"
setup-youtube-db = "src.scripts.setup_database:main"
backup-youtube-db = "src.scripts.backup_database:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]